Added initial comments, added Skybox, updated phong shader, added colour wheel ui
Added a skybox to the initial scene ensuring that it will not translate with the camera when it moves and will only rotate according to the cameras view matrix rotation. added colour wheel ui that allow you to change and edit the colour of certain objects. updated phong shader allowing it to use texture mapping.